// bounded-contexts/build-infra/domain/enums/google-resources.enum.ts
export enum GoogleResourcesEnum {
    VPC = "google_compute_network",
    SUB_VPC = "google_compute_subnetwork",
    FW = "google_compute_firewall",
    VM_GROUP = "google_compute_managed_instance_group",
    VM_TEMPLATE = "google_compute_instance_template",
    CACHE = "google_cloud_memorystore_instance",
    BUCKET = "google_storage_bucket",
    DB = "google_sql_database"
}